testscript: explore addition of io/fs.FS to Params
DO NOT SUBMIT
